Refactor: make d4ocr-rust can be used as a library.
